requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.12.5
Pillow==12.0.0
numpy==1.26.4
tensorflow==2.16.2
keras==3.12.0
h5py==3.15.1

